+++ itests/test_xss.py
@@ -2,18 +2,28 @@ import os
 
 import pytest
 import requests
+from _pytest.capture import CaptureFixture
 
 from keycloak_scanner.main import parser
 from keycloak_scanner.main import start
 
 
 @pytest.mark.skipif(os.getenv('ITESTS_XSS') != 'true', reason='integration tests')
-def test_should_start_scan_xss_fail_security_console_exit_4(base_url: str, capsys):
-
+def test_should_start_scan_xss_fail_security_console_exit_4(base_url: str, capsys: CaptureFixture, proxy: str):
     p = parser()
 
-    args = p.parse_args([base_url, '--realms', 'master', '--clients', 'account,account-console,admin-cli,broker,master-realm,security-admin-console',
-                         '--username', 'admin', '--password', 'Pa55w0rd'])
+    base_args = [base_url, '--realms', 'master', '--clients',
+                 'account,account-console,admin-cli,broker,master-realm,security-admin-console',
+                 '--username', 'admin', '--password', 'Pa55w0rd', '--registration-callback', 'http://localhost:8080']
+
+    if os.getenv('ITESTS_VERBOSE') == 'true':
+        base_args.append('--verbose')
+
+    if proxy:
+        base_args.append('--proxy')
+        base_args.append(proxy)
+
+    args = p.parse_args(base_args)
 
     with pytest.raises(SystemExit) as e:
         start(args, lambda: requests.Session())
@@ -26,7 +36,7 @@ def test_should_start_scan_xss_fail_security_console_exit_4(base_url: str, capsy
 
     print(captured.err)
 
-    assert captured.err == ''
+    assert captured.err == '[WARN] Result of ClientRegistrationScanner as no results (void list), subsequent scans can be void too.\n'
 
     assert 'Find realm master' in captured.out
 
@@ -38,8 +48,8 @@ def test_should_start_scan_xss_fail_security_console_exit_4(base_url: str, capsy
 
     assert "[INFO] Find a client auth endpoint for realm master: security-admin-console" in captured.out
 
-# TODO may be work after fix client scanner with registration
-#    assert "[+] LoginScanner - Form login work for admin on realm master, client security-admin-console" in captured.out
+    # TODO may be work after fix client scanner with registration
+    #    assert "[+] LoginScanner - Form login work for admin on realm master, client security-admin-console" in captured.out
 
     assert "[+] LoginScanner - Can login with username admin on realm master, client admin-cli, grant_type: password" in captured.out