+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - fail fast mode
 - Examples in doc
 - Don't display all objects on logs for better readability
+- client registration scanner
+- improve client scanner
+- callback list support for client registration scanner
+- delete client after client registration scan
 
 # version 1.0.3